python3 to make it run on raspberry

+++ src/steering_translator_node.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 import os
 
 import numpy as np
@@ -53,8 +53,8 @@ def handle_config_update(received_config, level):
 def handle_input(data):
     rospy.loginfo("new message {}".format(data))
     msg = pwm_steering()
-    msg.steering = np.interp(data.drive.steering_angle, INPUT_RANGE_STEERING, pwm_range_steering)
-    msg.throttle = np.interp(data.drive.speed, INPUT_RANGE_THROTTLE, pwm_range_throttle)
+    msg.steering = np.interp(data.drive.steering_angle, INPUT_RANGE_STEERING, pwm_range_steering).astype('int32')
+    msg.throttle = np.interp(data.drive.speed, INPUT_RANGE_THROTTLE, pwm_range_throttle).astype('int32')
     steering_publisher.publish(msg)
